rename config to _config
