+++ src/stpipe/_config.py
